Change type to `u32`

+++ src/model/application/command_interaction.rs
@@ -89,7 +89,7 @@ pub struct CommandInteraction {
     /// The context where the interaction was triggered from.
     pub context: Option<InteractionContext>,
     /// Attachment size limit in bytes
-    pub attachment_size_limit: u64,
+    pub attachment_size_limit: u32,
 }
 
 #[cfg(feature = "model")]
